etl/requirements.txt
clickhouse-driver==0.2.5
pydantic==1.10.5
python-dotenv==0.21.1
aiokafka==0.8.0